requirements.txt
Django==1.8
wheel==0.24.0
djangorestframework==3.4.0